phase-3-chatbot/backend/requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
sqlalchemy==1.4.46
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
psycopg2-binary==2.9.9
openai==1.54.0
mcp==1.1.0